Step 6 (complete).  Changed Album class:
Hid AlbumId field

Changed Display Names of:
GenreId, ArtistId and AlbumArtURL to Genre, Artist and Album Art URL

Required a title with a max length of 160 and put in an error message.

Required a price within parameters 0.01-100.00 with an error message to display.

Changed "public Genre Genre" to public virtual Genre Genre (applied to Artist as well).  To allow for EF to "lazy load" them as necessary.